Cloudant scripts.
Summary: Move cloudant db creation from the app code.

Test Plan: fullcycle, gradle run

Reviewers: txchen

Subscribers: all

Differential Revision: https://code.silkcloud.info/D313

+++ setupdb.sh
@@ -1,10 +1,22 @@
 #!/bin/sh
 source "$(git rev-parse --show-toplevel)/scripts/common.sh"; # this comment is needed, see common.sh for detail
 
-python ./scripts/couchdbcmd.py dropall --yes
-pushd liquibase
-./dropdb.sh
-./createdb.sh
-./updatedb.sh
-popd
+if [[ "$1" == "" || "$1" == "cloudant" ]]; then
+    # setup cloudant db
+    dbPrefixFile=common/configuration-data/src/main/resources/junbo/conf/onebox/common/personal.properties
+    dbPrefix=`cat $dbPrefixFile | grep '^common.cloudant.dbNamePrefix=' | sed 's/common.cloudant.dbNamePrefix=//'`
+    pushd cloudant
+    python ./couchdbcmd.py dropdbs --prefix=$dbPrefix --yes
+    python ./couchdbcmd.py createdbs --prefix=$dbPrefix --yes
+    popd
+fi
+
+if [[ "$1" == "" || "$1" == "sql" ]]; then
+    # setup sql db
+    pushd liquibase
+    ./dropdb.sh
+    ./createdb.sh
+    ./updatedb.sh
+    popd
+fi
